Html Writer Styles when Using Inline Css (#3680)
* Html Writer Styles when Using Inline Css

Fix #3678. Problem introduced by PR #3016. Combining `td` and `th` styles into a single declaration greatly reduces file size when `useInlineCss` is false, which is the default. However, generating code with the non-default option was not changed to use the combined declaration, so styling was lost. This PR rectifies that error.

* Apostrophe Rather Than Quote